Title: 합과 평균, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 231118/합과 평균/sum-and-avg.cpp
@@ -5,5 +5,5 @@ int main() {
     cout << a+b << " ";
     cout<<fixed;
     cout.precision(1);
-    cout << (a+b)/2;
+    cout << (double)(a+b)/2;
 }
